+++ app/src/main/java/com/techcourse/service/AppUserService.java
@@ -1,0 +1,64 @@
+package com.techcourse.service;
+
+import com.interface21.dao.DataAccessException;
+import com.interface21.jdbc.datasource.DataSourceUtils;
+import com.interface21.transaction.support.TransactionSynchronizationManager;
+import com.techcourse.config.DataSourceConfig;
+import com.techcourse.dao.UserDao;
+import com.techcourse.dao.UserHistoryDao;
+import com.techcourse.domain.User;
+import com.techcourse.domain.UserHistory;
+import java.sql.Connection;
+import java.sql.SQLException;
+import javax.sql.DataSource;
+
+public class AppUserService implements UserService {
+
+    private final UserDao userDao;
+    private final UserHistoryDao userHistoryDao;
+    private final DataSource dataSource;
+
+    public AppUserService(final UserDao userDao, final UserHistoryDao userHistoryDao) {
+        this.dataSource = DataSourceConfig.getInstance();
+        this.userDao = userDao;
+        this.userHistoryDao = userHistoryDao;
+    }
+
+    @Override
+    public User findById(final long id) {
+        return userDao.findById(id);
+    }
+
+    @Override
+    public void save(final User user) {
+        userDao.insert(user);
+    }
+
+    @Override
+    public void changePassword(final long id, final String newPassword, final String createBy) {
+        try {
+            // 트랜잭션 시작
+            Connection connection = DataSourceUtils.getConnection(dataSource);
+            connection.setAutoCommit(false);
+            try {
+                // 비즈니스 로직 처리
+                final var user = findById(id);
+                user.changePassword(newPassword);
+                userDao.update(connection, user);
+                userHistoryDao.log(connection, new UserHistory(user, createBy));
+                // 트랜잭션 커밋
+                connection.commit();
+            } catch (Exception e) {
+                // 트랜잭션 롤백
+                // 로직 처리 중에 예외가 발생하면 원자성을 보장하기 위해 롤백한다.
+                connection.rollback(); // try-catch로 한 번 더 감싸야 하지만 예시니까 생략
+                throw new DataAccessException(e);
+            } finally {
+                DataSourceUtils.releaseConnection(connection, dataSource);
+                TransactionSynchronizationManager.unbindResource(dataSource);
+            }
+        } catch (SQLException e) {
+            throw new DataAccessException(e);
+        }
+    }
+}
