+++ app/src/main/java/com/techcourse/service/TxUserService.java
@@ -1,0 +1,32 @@
+package com.techcourse.service;
+
+import com.techcourse.domain.User;
+
+public class TxUserService implements UserService {
+
+    private final UserService userService;
+
+    public TxUserService(final UserService userService) {
+        this.userService = userService;
+    }
+
+    @Override
+    public User findById(final long id) {
+        return userService.findById(id);
+    }
+
+    @Override
+    public void save(final User user) {
+        userService.save(user);
+    }
+
+    // override 대상인 메서드는 userService의 메서드를 그대로 위임(delegate)한다.
+    @Override
+    public void changePassword(final long id, final String newPassword, final String createdBy) {
+        // 트랜잭션 처리 영역
+
+        userService.changePassword(id, newPassword, createdBy);
+
+        // 트랜잭션 처리 영역
+    }
+}
